+++ cmd/rqlited/main.go
@@ -8,6 +8,7 @@ import (
 	"log"
 	"net"
 	"os"
+	"path/filepath"
 	"runtime"
 	"strings"
 	"syscall"
@@ -269,6 +270,23 @@ func createStore(cfg *Config, ln *tcp.Layer) (*store.Store, error) {
 	dbConf := store.NewDBConfig()
 	dbConf.OnDiskPath = cfg.OnDiskPath
 	dbConf.FKConstraints = cfg.FKConstraints
+
+	// Any SQLite Extensions to load?
+	if cfg.ExtensionsDir != "" {
+		files, err := os.ReadDir(cfg.ExtensionsDir)
+		if err != nil {
+			return nil, fmt.Errorf("failed to list files in Extensions directory: %s", err.Error())
+		}
+		for _, f := range files {
+			if !f.IsDir() && !strings.HasPrefix(f.Name(), ".") {
+				path := filepath.Join(cfg.ExtensionsDir, f.Name())
+				dbConf.Extensions = append(dbConf.Extensions, path)
+			}
+		}
+		if dbConf.Extensions != nil {
+			log.Printf("loading SQLite extensions: %v", dbConf.ExtensionNames())
+		}
+	}
 
 	str := store.New(ln, &store.Config{
 		DBConf: dbConf,